backend/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
python-dotenv
pydantic
email-validator
passlib[bcrypt]
python-jose
psycopg2-binary
bcrypt==4.0.1